refactor : 모호한 부분 재배치
