+++ src/main/java/kr/inuappcenterportal/inuportal/domain/fire/controller/FireController.java
@@ -13,8 +13,8 @@ import kr.inuappcenterportal.inuportal.domain.fire.dto.FirePageResponseDto;
 import kr.inuappcenterportal.inuportal.domain.fire.dto.FireRatingDto;
 import kr.inuappcenterportal.inuportal.domain.fire.dto.FireResponseDto;
 import kr.inuappcenterportal.inuportal.domain.member.model.Member;
-import kr.inuappcenterportal.inuportal.dto.*;
 import kr.inuappcenterportal.inuportal.domain.fire.service.FireService;
+import kr.inuappcenterportal.inuportal.global.dto.ResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
