+++ src/main/java/kr/inuappcenterportal/inuportal/domain/fire/model/Fire.java
@@ -1,7 +1,7 @@
 package kr.inuappcenterportal.inuportal.domain.fire.model;
 
 import jakarta.persistence.*;
-import kr.inuappcenterportal.inuportal.domain.BaseTimeEntity;
+import kr.inuappcenterportal.inuportal.global.model.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
