+++ src/main/java/kr/inuappcenterportal/inuportal/domain/folder/controller/FolderController.java
@@ -14,8 +14,9 @@ import kr.inuappcenterportal.inuportal.domain.folder.dto.FolderDto;
 import kr.inuappcenterportal.inuportal.domain.folder.dto.FolderPostDto;
 import kr.inuappcenterportal.inuportal.domain.folder.dto.FolderResponseDto;
 import kr.inuappcenterportal.inuportal.domain.member.model.Member;
-import kr.inuappcenterportal.inuportal.dto.*;
 import kr.inuappcenterportal.inuportal.domain.folder.service.FolderService;
+import kr.inuappcenterportal.inuportal.global.dto.ListResponseDto;
+import kr.inuappcenterportal.inuportal.global.dto.ResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
