+++ src/main/java/kr/inuappcenterportal/inuportal/domain/post/controller/PostController.java
@@ -15,9 +15,10 @@ import kr.inuappcenterportal.inuportal.domain.member.model.Member;
 import kr.inuappcenterportal.inuportal.domain.post.dto.PostDto;
 import kr.inuappcenterportal.inuportal.domain.post.dto.PostListResponseDto;
 import kr.inuappcenterportal.inuportal.domain.post.dto.PostResponseDto;
-import kr.inuappcenterportal.inuportal.dto.*;
 import kr.inuappcenterportal.inuportal.domain.post.service.PostService;
-import kr.inuappcenterportal.inuportal.service.RedisService;
+import kr.inuappcenterportal.inuportal.global.dto.ListResponseDto;
+import kr.inuappcenterportal.inuportal.global.dto.ResponseDto;
+import kr.inuappcenterportal.inuportal.global.service.RedisService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
