+++ src/main/java/kr/inuappcenterportal/inuportal/domain/reply/dto/ReReplyResponseDto.java
@@ -1,4 +1,4 @@
-package kr.inuappcenterportal.inuportal.dto;
+package kr.inuappcenterportal.inuportal.domain.reply.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import kr.inuappcenterportal.inuportal.domain.reply.model.Reply;
