+++ src/main/java/kr/inuappcenterportal/inuportal/domain/reply/service/ReplyService.java
@@ -4,7 +4,7 @@ import kr.inuappcenterportal.inuportal.domain.member.model.Member;
 import kr.inuappcenterportal.inuportal.domain.post.model.Post;
 import kr.inuappcenterportal.inuportal.domain.reply.model.Reply;
 import kr.inuappcenterportal.inuportal.domain.replylike.model.ReplyLike;
-import kr.inuappcenterportal.inuportal.dto.ReReplyResponseDto;
+import kr.inuappcenterportal.inuportal.domain.reply.dto.ReReplyResponseDto;
 import kr.inuappcenterportal.inuportal.domain.reply.dto.ReplyDto;
 import kr.inuappcenterportal.inuportal.domain.reply.dto.ReplyListResponseDto;
 import kr.inuappcenterportal.inuportal.domain.reply.dto.ReplyResponseDto;
@@ -13,7 +13,7 @@ import kr.inuappcenterportal.inuportal.global.exception.ex.MyException;
 import kr.inuappcenterportal.inuportal.domain.replylike.repository.LikeReplyRepository;
 import kr.inuappcenterportal.inuportal.domain.post.repository.PostRepository;
 import kr.inuappcenterportal.inuportal.domain.reply.repository.ReplyRepository;
-import kr.inuappcenterportal.inuportal.service.RedisService;
+import kr.inuappcenterportal.inuportal.global.service.RedisService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
