+++ src/main/java/kr/inuappcenterportal/inuportal/global/dto/ListResponseDto.java
@@ -1,4 +1,4 @@
-package kr.inuappcenterportal.inuportal.dto;
+package kr.inuappcenterportal.inuportal.global.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import kr.inuappcenterportal.inuportal.domain.post.dto.PostListResponseDto;
