+++ src/main/java/kr/inuappcenterportal/inuportal/global/exception/MyExceptionHandler.java
@@ -2,7 +2,7 @@ package kr.inuappcenterportal.inuportal.global.exception;
 
 
 import jakarta.validation.ConstraintViolationException;
-import kr.inuappcenterportal.inuportal.dto.ResponseDto;
+import kr.inuappcenterportal.inuportal.global.dto.ResponseDto;
 import kr.inuappcenterportal.inuportal.global.exception.ex.MyException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
