+++ src/main/java/kr/inuappcenterportal/inuportal/global/model/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package kr.inuappcenterportal.inuportal.domain;
+package kr.inuappcenterportal.inuportal.global.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
